+++ src/tooling/piral-cli/src/commands.ts
@@ -61,7 +61,7 @@ const allCommands: Array<ToolCommand<any>> = [
       return argv
         .positional('source', {
           type: 'string',
-          describe: 'Sets the source root directory or index.html file for collecting all the information.',
+          describe: 'Sets the source Piral instance path for collecting all the information.',
           default: apps.debugPiralDefaults.entry,
         })
         .string('target')
@@ -121,7 +121,7 @@ const allCommands: Array<ToolCommand<any>> = [
       return argv
         .positional('source', {
           type: 'string',
-          describe: 'Sets the source root directory or index.html file for collecting all the information.',
+          describe: 'Sets the source Piral instance path for collecting all the information.',
           default: apps.buildPiralDefaults.entry,
         })
         .string('target')
@@ -234,7 +234,7 @@ const allCommands: Array<ToolCommand<any>> = [
       return argv
         .positional('source', {
           type: 'string',
-          describe: 'Sets the source root directory or index.html file for collecting all the information.',
+          describe: 'Sets the source Piral instance path for collecting all the information.',
           default: apps.declarationPiralDefaults.entry,
         })
         .string('target')
@@ -388,7 +388,7 @@ const allCommands: Array<ToolCommand<any>> = [
       return argv
         .positional('source', {
           type: 'string',
-          describe: 'Sets the source root directory or index.html file for collecting all the information.',
+          describe: 'Sets the source Piral instance path for collecting all the information.',
           default: apps.validatePiralDefaults.entry,
         })
         .number('log-level')
@@ -489,7 +489,7 @@ const allCommands: Array<ToolCommand<any>> = [
       return argv
         .positional('source', {
           type: 'string',
-          describe: 'Sets the source index.tsx file for collecting all the information.',
+          describe: 'Sets the source pilet path for collecting all the information.',
           default: apps.buildPiletDefaults.entry,
         })
         .string('target')
@@ -568,7 +568,7 @@ const allCommands: Array<ToolCommand<any>> = [
       return argv
         .positional('source', {
           type: 'string',
-          describe: 'Sets the source package.json file for creating the package.',
+          describe: 'Sets the source pilet path for creating the package.',
           default: apps.packPiletDefaults.source,
         })
         .string('target')
@@ -804,6 +804,82 @@ const allCommands: Array<ToolCommand<any>> = [
         entry: args.entry as string,
         logLevel: args['log-level'] as LogLevels,
         app: args.app as string,
+      });
+    },
+  },
+  {
+    name: 'add-piral-instance-pilet',
+    alias: ['add-shell-pilet', 'install-shell-pilet', 'add-app-pilet', 'with-app-pilet'],
+    description: 'Adds another Piral instance to the available apps hosting the current pilet.',
+    arguments: ['<app> [source]'],
+    flags(argv) {
+      return argv
+        .positional('app', {
+          type: 'string',
+          describe: 'Sets the name of the Piral instance to be added.',
+          default: apps.addPiralInstancePiletDefaults.app,
+        })
+        .positional('source', {
+          type: 'string',
+          describe: 'Sets the source pilet path where the Piral instance should be added.',
+          default: apps.addPiralInstancePiletDefaults.source,
+        })
+        .number('log-level')
+        .describe('log-level', 'Sets the log level to use (1-5).')
+        .default('log-level', apps.addPiralInstancePiletDefaults.logLevel)
+        .choices('npm-client', clientTypeKeys)
+        .describe('npm-client', 'Sets the npm client to be used for adding the Piral instance.')
+        .default('npm-client', apps.addPiralInstancePiletDefaults.npmClient)
+        .boolean('selected')
+        .describe('selected', 'Defines if the provided Piral instance should be selected initially.')
+        .default('selected', apps.addPiralInstancePiletDefaults.selected)
+        .string('base')
+        .default('base', process.cwd())
+        .describe('base', 'Sets the base directory. By default the current directory is used.');
+    },
+    run(args) {
+      return apps.addPiralInstancePilet(args.base as string, {
+        logLevel: args['log-level'] as LogLevels,
+        selected: args.selected as boolean,
+        npmClient: args['npm-client'] as NpmClientType,
+        app: args.app as string,
+        source: args.source as string,
+      });
+    },
+  },
+  {
+    name: 'remove-piral-instance-pilet',
+    alias: ['remove-shell-pilet', 'uninstall-shell-pilet', 'remove-app-pilet', 'without-app-pilet'],
+    description: 'Removes an existing Piral instance from the available apps hosting the current pilet.',
+    arguments: ['<app> [source]'],
+    flags(argv) {
+      return argv
+        .positional('app', {
+          type: 'string',
+          describe: 'Sets the name of the Piral instance to be removed.',
+          default: apps.removePiralInstancePiletDefaults.app,
+        })
+        .positional('source', {
+          type: 'string',
+          describe: 'Sets the source pilet path where the Piral instance should be removed.',
+          default: apps.removePiralInstancePiletDefaults.source,
+        })
+        .number('log-level')
+        .describe('log-level', 'Sets the log level to use (1-5).')
+        .default('log-level', apps.removePiralInstancePiletDefaults.logLevel)
+        .choices('npm-client', clientTypeKeys)
+        .describe('npm-client', 'Sets the npm client to be used for removing the Piral instance.')
+        .default('npm-client', apps.removePiralInstancePiletDefaults.npmClient)
+        .string('base')
+        .default('base', process.cwd())
+        .describe('base', 'Sets the base directory. By default the current directory is used.');
+    },
+    run(args) {
+      return apps.removePiralInstancePilet(args.base as string, {
+        logLevel: args['log-level'] as LogLevels,
+        app: args.app as string,
+        source: args.source as string,
+        npmClient: args['npm-client'] as NpmClientType,
       });
     },
   },