+++ src/tooling/piral-cli/src/common/clients/pnp.ts
@@ -37,10 +37,17 @@ export async function installDependencies(target = '.', ...flags: Array<string>)
   return ms.value;
 }
 
+export async function uninstallPackage(packageRef: string, target = '.', ...flags: Array<string>) {
+  const ms = new MemoryStream();
+  await runYarnProcess(['remove', packageRef, ...convert(flags)], target, ms);
+  log('generalDebug_0003', `Yarn PnP remove package result: ${ms.value}`);
+  return ms.value;
+}
+
 export async function installPackage(packageRef: string, target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
   await runYarnProcess(['add', packageRef, ...convert(flags)], target, ms);
-  log('generalDebug_0003', `Yarn PnP install package result: ${ms.value}`);
+  log('generalDebug_0003', `Yarn PnP add package result: ${ms.value}`);
   return ms.value;
 }