+++ src/tooling/piral-cli/src/common/declaration.ts
@@ -2,7 +2,7 @@ import { DeclOptions, generateDeclaration, createExcludePlugin, Logger } from 'd
 import { dirname, basename, resolve, extname } from 'path';
 import { progress, log, logWarn, logVerbose, logInfo } from './log';
 import { ForceOverwrite } from './enums';
-import { retrievePiralRoot, retrievePiletsInfo } from './package';
+import { retrievePiralRoot, retrievePiletsInfo, flattenExternals } from './package';
 import { entryModuleExtensions, piralBaseRoot } from './constants';
 import { readText, getEntryFiles, matchFiles, createFileIfNotExists } from './io';
 import { getModulePath } from '../external';
@@ -167,7 +167,7 @@ export async function createPiralDeclaration(
     types: findDeclaredTypings(root),
     apis: findPiralBaseApi(root, framework),
     noModuleDeclaration: true,
-    imports: externals.map(m => m.name),
+    imports: flattenExternals(externals),
     logLevel,
     logger: createLogger(),
   };